code refactor\nsome logic updates
