+++ AmericaTVGoAuthProvider/Resources/AmericaTVGoIAPProductsViewController.xib
@@ -10,7 +10,7 @@
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
     </dependencies>
     <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="AmericaTVGoRegisterStep2PremiumViewController" customModule="AmericaTVGoAuthProvider" customModuleProvider="target">
+        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="AmericaTVGoIAPProductsViewController">
             <connections>
                 <outlet property="productsCollectionView" destination="f3z-Ez-WKs" id="MEN-IV-VDI"/>
                 <outlet property="productsCollectionViewHeightConstraint" destination="SCJ-l2-QHS" id="PJC-xn-dyW"/>
