+++ AmericaTVGoAuthProvider/Resources/AmericaTVGoRegisterAccountPickerViewController.xib
@@ -10,7 +10,7 @@
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
     </dependencies>
     <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="AmericaTVGoRegisterStep0ViewController" customModule="AmericaTVGoAuthProvider" customModuleProvider="target">
+        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="AmericaTVGoRegisterAccountPickerViewController">
             <connections>
                 <outlet property="basicBoxView" destination="NWM-JV-XKn" id="5Um-fl-tjH"/>
                 <outlet property="premiumBoxView" destination="jdg-qQ-Gfm" id="lak-S5-sRW"/>
